feat: Add support of the new chromedriver storage
