+++ lib/chromedriver.js
@@ -10,14 +10,13 @@ import {
   getChromedriverDir,
   CHROMEDRIVER_CHROME_MAPPING,
   getChromedriverBinaryPath,
-  CD_CDN,
   generateLogPrefix,
 } from './utils';
 import semver from 'semver';
 import _ from 'lodash';
 import path from 'path';
 import {compareVersions} from 'compare-versions';
-import ChromedriverStorageClient from './storage-client';
+import ChromedriverStorageClient from './storage-client/storage-client';
 import {toW3cCapNames, getCapValue} from './protocol-helpers';
 
 const NEW_CD_VERSION_FORMAT_MAJOR_VERSION = 73;
@@ -447,8 +446,7 @@ export class Chromedriver extends events.EventEmitter {
           } catch (e) {
             const err = /** @type {Error} */ (e);
             this.log.warn(
-              `Cannot synchronize local chromedrivers with the remote storage at ${CD_CDN}: ` +
-                err.message
+              `Cannot synchronize local chromedrivers with the remote storage: ${err.message}`
             );
             this.log.debug(err.stack);
           }
