+++ lib/install.js
@@ -1,11 +1,10 @@
 import { fs, mkdirp } from '@appium/support';
-import ChromedriverStorageClient from './storage-client';
+import ChromedriverStorageClient from './storage-client/storage-client';
 import {
-  CD_CDN, CD_VER, retrieveData, getOsInfo, getChromedriverDir,
+  CD_VER, retrieveData, getOsInfo, getChromedriverDir,
 } from './utils';
 
 
-const DOWNLOAD_TIMEOUT_MS = 15 * 1000;
 const LATEST_VERSION = 'LATEST';
 /**
  *
@@ -13,7 +12,7 @@ const LATEST_VERSION = 'LATEST';
  */
 async function formatCdVersion (ver) {
   return ver === LATEST_VERSION
-    ? (await retrieveData(`${CD_CDN}/LATEST_RELEASE`, {
+    ? (await retrieveData(`${GOOGLEAPIS_CDN}/LATEST_RELEASE`, {
       'user-agent': 'appium',
       accept: '*/*',
     }, { timeout: DOWNLOAD_TIMEOUT_MS })).trim()
