+++ lib/storage-client/storage-client.js
@@ -1,6 +1,5 @@
 import {
   getChromedriverDir,
-  CD_CDN,
   retrieveData,
   getOsInfo,
   OS,
@@ -8,16 +7,19 @@ import {
   X86,
   APPLE_ARM_SUFFIXES,
   convertToInt,
-} from './utils';
+} from '../utils';
 import _ from 'lodash';
-import xpath from 'xpath';
-import {DOMParser} from '@xmldom/xmldom';
 import B from 'bluebird';
 import path from 'path';
 import os from 'os';
 import {system, fs, logger, tempDir, zip, util, net} from '@appium/support';
+import {
+  STORAGE_REQ_TIMEOUT_MS,
+  doesCdNameMatchOsNameAndArchitecture,
+  GOOGLEAPIS_CDN,
+} from './utils';
+import {parseGoogleapiStorageXml} from './googleapis';
 
-const TIMEOUT_MS = 15000;
 const MAX_PARALLEL_DOWNLOADS = 5;
 
 const log = logger.getLogger('ChromedriverStorageClient');
@@ -33,180 +35,18 @@ async function isCrcOk(src, checksum) {
   return _.toLower(md5) === _.toLower(checksum);
 }
 
-/**
- *
- * @param {Node|Attr} parent
- * @param {string?} childName
- * @param {string?} text
- * @returns
- */
-function findChildNode(parent, childName = null, text = null) {
-  if (!childName && !text) {
-    return null;
-  }
-  if (!parent.hasChildNodes()) {
-    return null;
-  }
-
-  for (let childNodeIdx = 0; childNodeIdx < parent.childNodes.length; childNodeIdx++) {
-    const childNode = /** @type {Element|Attr} */ (parent.childNodes[childNodeIdx]);
-    if (childName && !text && childName === childNode.localName) {
-      return childNode;
-    }
-    if (text) {
-      const childText = extractNodeText(childNode);
-      if (!childText) {
-        continue;
-      }
-      if (childName && childName === childNode.localName && text === childText) {
-        return childNode;
-      }
-      if (!childName && text === childText) {
-        return childNode;
-      }
-    }
-  }
-  return null;
-}
-
-/**
- *
- * @param {Node?} node
- * @returns
- */
-function extractNodeText(node) {
-  return !node || !node.firstChild || !util.hasValue(node.firstChild.nodeValue)
-    ? null
-    : node.firstChild.nodeValue;
-}
 
 export class ChromedriverStorageClient {
   /**
    *
-   * @param {import('./types').ChromedriverStorageClientOpts} args
+   * @param {import('../types').ChromedriverStorageClientOpts} args
    */
   constructor(args = {}) {
-    const {chromedriverDir = getChromedriverDir(), timeout = TIMEOUT_MS} = args;
+    const {chromedriverDir = getChromedriverDir(), timeout = STORAGE_REQ_TIMEOUT_MS} = args;
     this.chromedriverDir = chromedriverDir;
     this.timeout = timeout;
     /** @type {ChromedriverDetailsMapping} */
     this.mapping = {};
-  }
-
-  /**
-   * Gets additional chromedriver details from chromedriver
-   * release notes
-   *
-   * @param {string} content - Release notes of the corresponding chromedriver
-   * @returns {import('./types').AdditionalDriverDetails}
-   */
-  parseNotes(content) {
-    const result = {};
-    const versionMatch = /^\s*[-]+ChromeDriver[\D]+([\d.]+)/im.exec(content);
-    if (versionMatch) {
-      result.version = versionMatch[1];
-    }
-    const minBrowserVersionMatch = /^\s*Supports Chrome[\D]+(\d+)/im.exec(content);
-    if (minBrowserVersionMatch) {
-      result.minBrowserVersion = minBrowserVersionMatch[1];
-    }
-    return result;
-  }
-
-  /**
-   * Downloads chromedriver release notes and puts them
-   * into the dictionary argument
-   *
-   * The method call mutates by merging `AdditionalDriverDetails`
-   * @param {string} driverKey - Driver version plus archive name
-   * @param {string} notesUrl - The URL of chromedriver notes
-   * @param {ChromedriverDetails} infoDict - The dictionary containing driver info.
-   * @throws {Error} if the release notes cannot be downloaded
-   */
-  async retrieveAdditionalDriverInfo(driverKey, notesUrl, infoDict) {
-    const notes = await retrieveData(
-      notesUrl,
-      {
-        'user-agent': 'appium',
-        accept: '*/*',
-      },
-      {timeout: this.timeout}
-    );
-    const {minBrowserVersion} = this.parseNotes(notes);
-    if (!minBrowserVersion) {
-      log.debug(
-        `The driver '${driverKey}' does not contain valid release notes at ${notesUrl}. ` +
-          `Skipping it`
-      );
-      return;
-    }
-    infoDict.minBrowserVersion = minBrowserVersion;
-  }
-
-  /**
-   * Parses chromedriver storage XML and stores
-   * the parsed results into `this.mapping`
-   *
-   * @param {Document} doc - The DOM representation
-   * of the chromedriver storage XML
-   * @param {boolean} shouldParseNotes [true] - If set to `true`
-   * then additional drivers information is going to be parsed
-   * and assigned to `this.mapping`
-   */
-  async parseStorageXml(doc, shouldParseNotes = true) {
-    const driverNodes = /** @type {Array<Node|Attr>} */ (
-      xpath.select(`//*[local-name(.)='Contents']`, doc)
-    );
-    log.debug(`Parsed ${driverNodes.length} entries from storage XML`);
-    if (_.isEmpty(driverNodes)) {
-      return;
-    }
-
-    const promises = [];
-    for (const driverNode of driverNodes) {
-      const k = extractNodeText(findChildNode(driverNode, 'Key'));
-      if (!_.includes(k, '/chromedriver_')) {
-        continue;
-      }
-      const key = String(k);
-
-      const etag = extractNodeText(findChildNode(driverNode, 'ETag'));
-      if (!etag) {
-        log.debug(`The entry '${key}' does not contain the checksum. Skipping it`);
-        continue;
-      }
-
-      /** @type {ChromedriverDetails} */
-      const cdInfo = {
-        url: `${CD_CDN}/${key}`,
-        etag: _.trim(etag, '"'),
-        version: /** @type {string} */ (_.first(key.split('/'))),
-        minBrowserVersion: null,
-      };
-      this.mapping[key] = cdInfo;
-
-      const notesPath = `${cdInfo.version}/notes.txt`;
-      const isNotesPresent = !!driverNodes.reduce(
-        (acc, node) => Boolean(acc || findChildNode(node, 'Key', notesPath)),
-        false
-      );
-      if (!isNotesPresent) {
-        cdInfo.minBrowserVersion = null;
-        if (shouldParseNotes) {
-          log.info(`The entry '${key}' does not contain any notes. Skipping it`);
-        }
-        continue;
-      } else if (!shouldParseNotes) {
-        continue;
-      }
-
-      promises.push(this.retrieveAdditionalDriverInfo(key, `${CD_CDN}/${notesPath}`, cdInfo));
-      if (promises.length % MAX_PARALLEL_DOWNLOADS === 0) {
-        await B.all(promises);
-      }
-    }
-    await B.all(promises);
-    log.info(`The total count of entries in the mapping: ${_.size(this.mapping)}`);
   }
 
   /**
@@ -219,16 +59,15 @@ export class ChromedriverStorageClient {
    */
   async retrieveMapping(shouldParseNotes = true) {
     const xml = await retrieveData(
-      CD_CDN,
+      GOOGLEAPIS_CDN,
       {
         'user-agent': 'appium',
         accept: 'application/xml, */*',
       },
       {timeout: this.timeout}
     );
-    const doc = new DOMParser().parseFromString(xml);
-    await this.parseStorageXml(doc, shouldParseNotes);
-    return _.cloneDeep(this.mapping);
+    this.mapping = await parseGoogleapiStorageXml(xml, shouldParseNotes);
+    return this.mapping;
   }
 
   /**
@@ -329,7 +168,9 @@ export class ChromedriverStorageClient {
     if (!_.isEmpty(osInfo)) {
       // Filter out drivers for unsupported system architectures
       let {name, arch} = osInfo;
-      if (arch === X64 && !driversToSync.some((cdName) => cdName.includes(`_${name}${X64}`))) {
+      if (arch === X64 && !driversToSync.some(
+        (cdName) => doesCdNameMatchOsNameAndArchitecture(cdName, name, X64)
+      )) {
         // Fall back to x86 build if x64 one is not available for the given OS
         arch = X86;
       }
@@ -344,8 +185,7 @@ export class ChromedriverStorageClient {
         }
       }
       log.debug(`Selecting chromedrivers whose platform matches to ${name}${arch}`);
-      const platformRe = new RegExp(`(\\b|_)${name}${arch}\\b`);
-      driversToSync = driversToSync.filter((cdName) => platformRe.test(cdName));
+      driversToSync = driversToSync.filter((cdName) => doesCdNameMatchOsNameAndArchitecture(cdName, name, arch));
       log.debug(`Got ${util.pluralize('item', driversToSync.length, true)}`);
     }
 
@@ -374,7 +214,7 @@ export class ChromedriverStorageClient {
     try {
       await net.downloadFile(url, archivePath, {
         isMetered: false,
-        timeout: TIMEOUT_MS,
+        timeout: STORAGE_REQ_TIMEOUT_MS,
       });
     } catch (e) {
       const err = /** @type {Error} */ (e);
@@ -464,7 +304,7 @@ export class ChromedriverStorageClient {
     if (!_.isEmpty(synchronizedDrivers)) {
       log.info(
         `Successfully synchronized ` +
-          `${util.pluralize('chromedriver', synchronizedDrivers.length, true)}`
+        `${util.pluralize('chromedriver', synchronizedDrivers.length, true)}`
       );
     } else {
       log.info(`No chromedrivers were synchronized`);
@@ -476,8 +316,8 @@ export class ChromedriverStorageClient {
 export default ChromedriverStorageClient;
 
 /**
- * @typedef {import('./types').SyncOptions} SyncOptions
- * @typedef {import('./types').OSInfo} OSInfo
- * @typedef {import('./types').ChromedriverDetails} ChromedriverDetails
- * @typedef {import('./types').ChromedriverDetailsMapping} ChromedriverDetailsMapping
+ * @typedef {import('../types').SyncOptions} SyncOptions
+ * @typedef {import('../types').OSInfo} OSInfo
+ * @typedef {import('../types').ChromedriverDetails} ChromedriverDetails
+ * @typedef {import('../types').ChromedriverDetailsMapping} ChromedriverDetailsMapping
  */
