+++ lib/utils.js
@@ -5,10 +5,6 @@ import path from 'path';
 import {compareVersions} from 'compare-versions';
 import axios from 'axios';
 
-const CD_CDN =
-  process.env.npm_config_chromedriver_cdnurl ||
-  process.env.CHROMEDRIVER_CDNURL ||
-  'https://chromedriver.storage.googleapis.com';
 const OS = {
   linux: 'linux',
   windows: 'win',
@@ -175,7 +171,6 @@ export {
   getChromedriverBinaryPath,
   getOsName,
   CD_BASE_DIR,
-  CD_CDN,
   CD_VER,
   CHROMEDRIVER_CHROME_MAPPING,
   getMostRecentChromedriver,
