+++ test/functional/storage-client-e2e-specs.js
@@ -2,7 +2,7 @@
 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import ChromedriverStorageClient from '../../lib/storage-client';
+import ChromedriverStorageClient from '../../lib/storage-client/storage-client';
 import _ from 'lodash';
 import { fs, tempDir } from '@appium/support';
